Safari-609.1.11.1/LayoutTests/css3/flexbox/flex-align-percent-height.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
body {
    margin: 0;
}

body, html {
    height: 600px;
}

.flexbox {
    width: 600px;
    background-color: #aaa;
    position: relative;
}

.flexbox div {
    border: 0;
}

.flexbox :nth-child(1) {
    background-color: blue;
}
.flexbox :nth-child(2) {
    background-color: green;
}

</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>

<div class="flexbox" style="height: 50%;">
  <div data-expected-height="50"  data-offset-y="0" class="flex-one" style="height: 50px;"></div>
  <div data-expected-height="300" data-offset-y="0" class="flex-one"></div>
</div>

</body>
</html>
